taller_mecanico/requirements.txt
amqp==5.3.1
APScheduler==3.11.2
asgiref==3.8.1
billiard==4.2.4
celery==5.6.3
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.4
Django==5.2.1
django-apscheduler==0.7.0
django-crispy-forms==2.4
django-extensions==4.1
django_celery_results==2.6.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
kombu==5.6.2
mysqlclient==2.2.7
packaging==26.0
pillow==11.2.1
prompt_toolkit==3.0.52
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
